WIP: Debugging info for null transactions
We're seeing some null txs, and need more information to diagnose.

This update will get transactions from disk if they aren't available
from the mempool.

If a transaction is known to the stream but available from neither the
mempool nor the disk, it will provide a note with the hash of the
transaction.

There is also a "note" field which indicates whether the transaction
was found in the mempool, disk, or if it was not found the hash of
the transaction.

This is not intended to be merged for production use.