Fix null interface issue
I inadvertently got tripped up by the way Go handles null interfaces.
If you say:

    var value interface{}
    value = tx

and tx is a nil pointer to a transaction, then value != nil
because it is not an empty interface, even though tx == nil.

It gets very confusing when this statement can be true:

    tx == value && tx == nil && value != nil

but this was what was going on here. This takes out the value
entirely, since we're capturing the hash in the note field when
needed.