Eliminate concurrent map access, add channel buffer
While testing looking for a lockup that would prevent blocks / transactions
from processing, I ran across a concurrent map access that panicked Geth.
The lock should fix the concurrent map access.

I also added buffer to the channel for receiving transactions. A client
that is falling behind in its processing of messages could have caused
that goroutine to slow, causing upstream goroutines to block waiting for
it to catch up with the channel. I don't have clear evidence of this
happening, but it's the most plausible explanation for the behavior
described.